// types.ts
export type Language = 'en' | 'th';

export type LocalizedString = {
  en: string;
  th: string;
};
// ...
export enum ProductStatus {
  IN_STOCK = 'In Stock',
  LOW_STOCK = 'Low Stock',
  OUT_OF_STOCK = 'Out of Stock',
}

export interface StockHistory {
    date: string;
    change: number;
    reason: string;
    operator: string;
}

export interface ProductPrice {
  walkIn: number;
  contractor: number;
  government: number;
  cost: number;
}

export interface ProductVariant {
  id: string;
  sku: string;
  size: string;
  stock: number;
  price: ProductPrice;
  status: ProductStatus;
  barcode?: string;
  history: StockHistory[];
}

export type NewProductVariantData = Omit<ProductVariant, 'id' | 'status' | 'history'>;

export interface Product {
  id: string;
  name: LocalizedString;
  description?: LocalizedString;
  category: string; // e.g., 'building_materials.cement_aggregates'
  imageUrl: string;
  variants: ProductVariant[];
}